Updated build.gradle to be cleaner: - Added Fastboard through Gradle - Changed most implementations to libraries in plugin.yml (lighter plugin jar) except Fastboard which needs relocation - Removed old Fastboard classes - Changed the imports for Fastboard to use the gradle library - Changed Paper import from manual to paperweight to support NMS and mappings for them
