+++ build.gradle
@@ -1,88 +1,62 @@
 plugins {
-    id 'com.github.johnrengelman.shadow' version '8.1.1'
     id 'java'
+
+    id "io.papermc.paperweight.userdev" version "1.7.2"
+    id "io.freefair.lombok" version "8.10"
+    id 'com.gradleup.shadow' version '8.3.0'
 }
 
 group = 'fr.tetemh'
 version = "0.1"
 
-apply plugin: 'java'
 compileJava.options.encoding = 'UTF-8'
-sourceCompatibility = targetCompatibility = JavaVersion.VERSION_21
-jar.getArchiveFileName().set("${project.name}.jar")
-libsDirName = "${jarOutput}"
 
 repositories {
     mavenCentral()
-    mavenLocal ()
-    maven {
-        url = "https://arcanearts.jfrog.io/artifactory/archives"
-        url = 'https://jitpack.io'
-        url = 'https://oss.sonatype.org/content/repositories/snapshots'
-        url = 'https://oss.sonatype.org/content/repositories/central'
-        url = "https://repo.papermc.io/repository/maven-public/"
-        url = "https://repo.fancyplugins.de/releases"
-    }
 
+    maven { url = 'https://jitpack.io' }
+    maven { url = "https://repo.fancyplugins.de/releases" }
 }
 
 
 dependencies {
+    // JetBrains Annotations
+    compileOnly 'org.jetbrains:annotations:20.1.0'
+
+    // Paper
+    paperweight.paperDevBundle('1.21.1-R0.1-SNAPSHOT')
+
+    // Mojang Authlib
     compileOnly 'com.mojang:authlib:1.5.25'
 
-    implementation 'org.jetbrains:annotations:20.1.0'
-
-    // paper
-    compileOnly "io.papermc.paper:paper-api:1.21.1-R0.1-SNAPSHOT"
-    implementation group: 'net.kyori', name: 'adventure-api', version: '4.17.0'
-
-    // lombok
-    compileOnly 'org.projectlombok:lombok:1.18.30'
-    annotationProcessor 'org.projectlombok:lombok:1.18.30'
-
     // Fancy NPC
+    compileOnly 'de.oliver:FancyNpcs:2.2.2'
 
-    compileOnly("de.oliver:FancyNpcs:2.2.2")
+    // Fastboard
+    implementation 'fr.mrmicky:fastboard:2.1.3'
 
-    //bdd
-    implementation 'com.zaxxer:HikariCP:4.0.3'
-    implementation group: 'org.mariadb.jdbc', name: 'mariadb-java-client', version: '3.0.8'
-}
-
-sourceSets {
-    configurations.compileOnly.setCanBeResolved(true)
+    // DB
+    compileOnly 'com.zaxxer:HikariCP:4.0.3'
+    compileOnly 'org.mariadb.jdbc:mariadb-java-client:3.0.8'
 }
 
 def targetJavaVersion = 21
 java {
     def javaVersion = JavaVersion.toVersion(targetJavaVersion)
+
     sourceCompatibility = javaVersion
     targetCompatibility = javaVersion
-    if (JavaVersion.current() < javaVersion) {
-        toolchain.languageVersion = JavaLanguageVersion.of(targetJavaVersion)
-    }
+
+    toolchain.languageVersion = JavaLanguageVersion.of(targetJavaVersion)
 }
 
-tasks.withType(JavaCompile).configureEach {
-    if (targetJavaVersion >= 10 || JavaVersion.current().isJava10Compatible()) {
-        options.release = targetJavaVersion
-    }
-}
-
-processResources {
-    def props = [version: version]
-    inputs.properties props
-    filteringCharset 'UTF-8'
-    filesMatching('plugin.yml') {
-        expand props
-    }
-}
-
-artifacts {
-    archives shadowJar
+assemble {
+    dependsOn(tasks.reobfJar)
 }
 
 shadowJar {
-    relocate 'fr.mrmicky.fastboard', 'fr.tetemh.skycite.fastboard'
-}
+    relocate 'fr.mrmicky.fastboard', 'fr.tetemh.fastboard'
 
+    archiveFileName = "${project.name}.jar"
+    destinationDirectory = file(jarOutput)
+}
